+++ toolkit/scalebar/src/main/java/com/arcgismaps/toolkit/scalebar/ScalebarUnits.kt
@@ -37,4 +37,3 @@ public enum class ScalebarUnits {
      */
     METRIC
 }
-
