+++ toolkit/scalebar/src/main/java/com/arcgismaps/toolkit/scalebar/internal/ScalebarRenderer.kt
@@ -14,115 +14,31 @@
  * limitations under the License.
  */
 
-package com.arcgismaps.toolkit.scalebar
+package com.arcgismaps.toolkit.scalebar.internal
 
-import android.content.Context
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.width
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shadow
 import androidx.compose.ui.graphics.drawscope.DrawScope
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.TextMeasurer
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.drawText
-import androidx.compose.ui.text.rememberTextMeasurer
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.sp
 
-private const val pixelAlignment = 2.5f // Aligns the horizontal line edges
+internal const val pixelAlignment = 2.5f // Aligns the horizontal line edges
 private const val lineWidth = 5f
-private const val shadowOffset = 3f
-private const val scalebarHeight = 20f // Height of the scalebar in pixels
-private val textSize = 14.sp
-private const val textOffset = 5f
-
-/**
- * A composable UI component to display a line scalebar.
- * 
- * @since 200.7.0
- */
-@Composable
-internal fun LineScalebarImpl(
-    modifier: Modifier = Modifier,
-    scaleValue: String,
-    width: Float,
-    lineColor: Color = Color.Black,
-    shadowColor: Color = Color.Unspecified,
-    textColor: Color = Color.Black,
-    textShadowColor: Color = Color.Unspecified
-) {
-    val textMeasurer = rememberTextMeasurer()
-    val density = LocalDensity.current
-    val texSizeInPx = with(density) { textSize.toPx() }
-
-    val totalHeight = scalebarHeight + shadowOffset + textOffset + texSizeInPx
-    val totalWidth = width + shadowOffset + pixelAlignment
-
-    Canvas(
-        modifier = modifier
-            .width(calculateSizeInDp(density, totalWidth))
-            .height(calculateSizeInDp(density, totalHeight))
-    ) {
-        // left line
-        drawVerticalLine(
-            x = 0f,
-            top = 0f,
-            bottom = scalebarHeight,
-            color = lineColor,
-            shadowColor = shadowColor
-        )
-
-        // bottom line
-        drawHorizontalLine(
-            y = scalebarHeight,
-            left = 0f,
-            right = width,
-            color = lineColor,
-            shadowColor = shadowColor,
-        )
-
-        // right line
-        drawVerticalLine(
-            x = width,
-            top = 0f,
-            bottom = scalebarHeight,
-            color = lineColor,
-            shadowColor = shadowColor,
-        )
-        // text label
-        drawText(
-            text = scaleValue,
-            textMeasurer = textMeasurer,
-            barEnd = width,
-            scalebarHeight = scalebarHeight,
-            color = textColor,
-            shadowColor = textShadowColor,
-            alignment = TextAlignment.CENTER
-        )
-    }
-}
-
-@Composable
-internal fun isMetric(): Boolean {
-    // TODO implement the actual logic to determine the default ScalebarUnit
-    // this is a placeholder implementation
-    val context = LocalContext.current
-    val sharedPreferences = context.getSharedPreferences("settings", Context.MODE_PRIVATE)
-    return sharedPreferences.getBoolean("isMetric", true)
-}
+internal const val shadowOffset = 3f
+internal const val scalebarHeight = 20f // Height of the scalebar in pixels
+internal val textSize = 14.sp
+internal const val textOffset = 5f
 
 /**
  * Calculates the size in dp based on the density of the device.
  *
  * @since 200.7.0
  */
-private fun calculateSizeInDp(density: Density, value: Float) = with(density) {
+internal fun calculateSizeInDp(density: Density, value: Float) = with(density) {
     value.toDp()
 }
 
@@ -226,4 +142,3 @@ internal fun DrawScope.drawText(
         shadow = Shadow(color = shadowColor, offset = Offset(1f, 1f))
     )
 }
-
